refactor(basic): basic auth 패키지 명확하게 변경
