+++ backend/src/test/java/codezap/auth/provider/basic/BasicAuthDecoderTest.java
@@ -1,4 +1,4 @@
-package codezap.auth;
+package codezap.auth.provider.basic;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
